refactor: Move KrTabsBorder to a separate package and make it abstract.
