+++ src/main/java/krasa/editorGroups/tabs2/impl/border/KrTabsBorder.kt
@@ -1,4 +1,4 @@
-package krasa.editorGroups.tabs2
+package krasa.editorGroups.tabs2.impl.border
 
 import com.intellij.util.ui.JBUI
 import krasa.editorGroups.tabs2.impl.KrTabsImpl
